Add numpy recarray format to ps1cone query
The current formats for ps1cone are csv, votable, and json. This update allows the user to receive their results as a numpy recarray. The query requests a csv from the server and then converts that csv result into a recarray using `csvtonumpy`. This function converts the csv results by running a `ps1metadata` query for the same table and release, and then using the values that function returns along with the csv results the header row to create the dtype for the recarray. The rest of the csv results are turned into the data of the recarray.

This commit also removes extraneous imports.