Give explicit control over pagesize of results
Pagesize controls the number of maximum records that can be executed in a query. This script, as currently written, has a hidden limit of 50,000 maximum entries returned. This commit adds a pagesize parameter that defaults to 0, or no limit on the maximum entries returned. The user is also able to set this limit if desired.